Refactored and restructured playground scripts
