+++ .parcelrc
@@ -1,5 +1,6 @@
 {
   "extends": "@parcel/config-default",
+  "bundler": "@parcel/bundler-experimental",
   "optimizers": {
       "*.css": [],
   },
