+++ src/partials/default-body-top.ejs
@@ -1,6 +1,6 @@
 <script defer src="<%= rootDir %>/js/plugins.ts"></script>
 <script defer src="<%= rootDir %>/js/markdown.ts"></script>
-<script defer src="<%= rootDir %>/prism/prism.js"></script>
+<script defer src="<%= rootDir %>/prism/prism-core.js"></script>
 
 <!-- Show Error if JS is disabled or blocked -->
 <noscript>Oops that shouldn't have happened! Please enable Java-Script to use our site!</noscript>
