+++ src/partials/meta-head.ejs
@@ -29,7 +29,7 @@ let displayTitle = (typeof dropdownTitle !== "undefined" ? `${dropdownTitle}/` :
 <link rel="shortcut icon" type="image/ico" href="../<%= rootDir %>/favicon.ico" />
 
 <!-- Loading the prism highlighting style-sheet -->
-<link rel="stylesheet" href="<%= rootDir %>/prism/prism.css" />
+<link rel="stylesheet" href="<%= rootDir %>/prism/prism-core.css" />
 
 <!-- Font Awesome Icons -->
 <script src="https://kit.fontawesome.com/951ef1f046.js" crossorigin="anonymous"></script>
